Headers, coding, POD
Added tests for headers: checks the headers of the file against the headers in the specification and vice versa
Changed a lot of the coding. All the tests are now written in test_* while the asserts test a value against expected. This should improve maintainability a lot.
Updated the POD with the new methods, added headers for the methods, fixed a bug in the synopsis.